utils(commands): move mult replace from file
